Moved DefaultSelectionsRequest and DefaultSelectionsResponse classes to package message
